refactor token tests
